22/02/2024 - cambios en el slider

+++ app/src/main/java/com/example/ejerciciosexoplayer/screens/PlayerScreen.kt
@@ -76,7 +76,7 @@ fun ExoPlayerScreen(viewModelScaffold: ScaffoldViewModel = viewModel()){
             var valor = (posicion/1000).toFloat()
             Slider(
                 value = valor,
-                onValueChange = { exoPlayerViewModel.CambiarProgreso(it.toInt())},
+                onValueChange = { exoPlayerViewModel.CambiarProgreso((it).toLong())},
                 colors = SliderDefaults.colors(
                     thumbColor = MaterialTheme.colorScheme.secondary,
                     activeTrackColor = MaterialTheme.colorScheme.secondary,
